+++ CHANGELOG.md
@@ -20,6 +20,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
   Users who are using Kludgine directly should invoke `Kludgine::resize` to
   apply the new size.
+- `Window::winit` now returns an `Arc`-wrapped winit Window.
 
 ### Added
 
@@ -29,6 +30,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - `WindowBehavior::unrecoverable_error` is a new function that allows
   applications to take control when an unrecoverable error occurs. Previously,
   these errors would panic, and the provided implementation is a panic.
+- `App::execute` executes a closure on the main event loop thread.
 
 ## v0.11.0 (2024-09-14)